refactor: db테이블 자료형 변경으로 인한 수정(칼로리, 나트륨)

+++ src/main/java/xyz/teamcarrot/myct/healthInfo/FoodDiaryInfoVO.java
@@ -6,11 +6,11 @@ import lombok.Data;
 public class FoodDiaryInfoVO {
 	private int food_diary_info_no;
 	private String food_name;
-	private double food_calorie;
+	private int food_calorie;
 	private double food_carbs;
 	private double food_protein;
 	private double food_fat;
 	private double food_sugar;
-	private double food_salt;
+	private int food_salt;
 	private int food_diary_no;
 }
